+++ services/web/server/tests/unit/conftest.py
@@ -13,8 +13,11 @@ import logging
 import sys
 from pathlib import Path
 from typing import Dict
+from uuid import uuid4
 
 import pytest
+
+from simcore_service_webserver.utils import now_str
 
 ## current directory
 current_dir = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
@@ -39,6 +42,22 @@ def fake_static_dir(fake_data_dir: Path) -> Dict:
 def fake_project(fake_data_dir: Path) -> Dict:
     with (fake_data_dir / "fake-project.json").open() as fp:
         yield json.load(fp)
+
+@pytest.fixture
+def empty_project():
+    def create():
+        empty_project = {
+            "uuid": f"project-{uuid4()}",
+            "name": "Empty name",
+            "description": "some description of an empty project",
+            "prjOwner": "I'm the empty project owner, hi!",
+            "creationDate": now_str(),
+            "lastChangeDate": now_str(),
+            "thumbnail": "",
+            "workbench": {}
+        }
+        return empty_project
+    return create
 
 
 @pytest.fixture
